refactor(compaction): extract openai remote conversion into openai-remote-convert module
Pure move: wire item types, details schema guard, and branch-to-input
conversion now live in openai-remote-convert.ts. No behavior change;
openai-remote.ts re-exports the moved public symbols. Existing suite:
22 files / 170 tests green, tsgo + biome clean.
